+++ src/interfaces/index.ts
@@ -7,6 +7,7 @@ export * from './debounce-options';
 export * from './flow-job';
 export * from './ioredis-events';
 export * from './job-json';
+export * from './job-scheduler-json';
 export * from './keep-jobs';
 export * from './metrics-options';
 export * from './metrics';