+++ tests/test_repeat.ts
@@ -45,7 +45,7 @@ describe('repeat', function () {
   });
 
   beforeEach(async function () {
-    this.clock = sinon.useFakeTimers();
+    this.clock = sinon.useFakeTimers({ shouldClearNativeTimers: true });
     queueName = `test-${v4()}`;
     queue = new Queue(queueName, { connection, prefix });
     repeat = new Repeat(queueName, { connection, prefix });
@@ -1110,7 +1110,7 @@ describe('repeat', function () {
   });
 
   it('should repeat 7:th day every month at 9:25', async function () {
-    this.timeout(15000);
+    this.timeout(12000);
 
     const date = new Date('2017-02-02 7:21:42');
     this.clock.setSystemTime(date);
@@ -1162,7 +1162,7 @@ describe('repeat', function () {
     await queue.add(
       'repeat',
       { foo: 'bar' },
-      { repeat: { pattern: '* 25 9 7 * *' } },
+      { repeat: { pattern: '25 9 7 * *' } },
     );
     nextTick();